#315 move external interactions to submodule
