+++ src/artemis/external_interaction/nexus_writing/tests/test_data/dummy_0_000002.h5
